content/media/video/test/test_bug500311.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=500311
-->

<head>
  <title>Media test: Bug 500311 - Crash loading certain Ogg Theora files</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=500311">Mozilla Bug 500311</a>
<pre id="test">
<script class="testbody" type="text/javascript">
function endTest() {
  var v = document.getElementById("v");
  ok(v.duration > 1.9 && v.duration < 2.1, "Duration should be around 1.96");
  ok(v.ended, "Checking playback has ended");
  SimpleTest.finish();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<video id='v'
       src='bug500311.ogv'
       autoplay
       onended='return endTest();'></video>
</body>
</html>
